Removendo class 'pessoa', recriando o bando de dados
