--- src/main/java/com/br/lymtt/easypass/controller/Pessoacontroller.java
@@ -1,5 +1,0 @@
-package com.br.lymtt.easypass.controller;
-
-public class Pessoacontroller {
-    
-}
