+++ src/main/java/com/br/lymtt/easypass/model/entities/Cliente.java
@@ -1,7 +1,23 @@
 package com.br.lymtt.easypass.model.entities;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 
-public class Cliente extends Pessoa {
+@Entity
+@Table(schema = "LYMTT", name = "cliente")
 
+public class Cliente {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private String nome;
+    private String cpf;
+    private String genero; // masculino ou feminino
+    private int idade;
     private String email;
     private String numeroTelefone;
 
@@ -9,9 +25,53 @@ public class Cliente extends Pessoa {
     }
 
     public Cliente(Long id, String nome, String cpf, String genero, int idade, String email, String numeroTelefone) {
-        super(id, nome, cpf, genero, idade);
+        this.id = id;
+        this.nome = nome;
+        this.cpf = cpf;
+        this.genero = genero;
+        this.idade = idade;
         this.email = email;
         this.numeroTelefone = numeroTelefone;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getNome() {
+        return nome;
+    }
+
+    public void setNome(String nome) {
+        this.nome = nome;
+    }
+
+    public String getCpf() {
+        return cpf;
+    }
+
+    public void setCpf(String cpf) {
+        this.cpf = cpf;
+    }
+
+    public String getGenero() {
+        return genero;
+    }
+
+    public void setGenero(String genero) {
+        this.genero = genero;
+    }
+
+    public int getIdade() {
+        return idade;
+    }
+
+    public void setIdade(int idade) {
+        this.idade = idade;
     }
 
     public String getEmail() {
@@ -32,7 +92,8 @@ public class Cliente extends Pessoa {
 
     @Override
     public String toString() {
-        return "Cliente [email=" + email + ", numeroTelefone=" + numeroTelefone + ", id=" + getId() + ", nome=" + getNome() + ", cpf=" + getCpf() + ", genero=" + getGenero() + ", idade=" + getIdade() + "]";
+        return "Cliente [email=" + getEmail() + ", numeroTelefone=" + getNumeroTelefone() + ", id=" + getId() + ", nome="
+                + getNome() + ", cpf=" + getCpf() + ", genero=" + getGenero() + ", idade=" + getIdade() + "]";
     }
 
 }
