--- src/main/java/com/br/lymtt/easypass/model/repository/PessoaRepository.java
@@ -1,5 +1,0 @@
-package com.br.lymtt.easypass.model.repository;
-
-public class PessoaRepository {
-    
-}
